requirements.txt
alabaster==0.7.16
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
deepmerge==1.1.1
docutils==0.21.1
furo==2024.1.29
idna==3.7
imagesize==1.4.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
latexcodec==3.0.0
MarkupSafe==2.1.5
mistune==2.0.5
packaging==24.0
picobox==4.0.0
pybtex==0.24.0
pybtex-docutils==1.0.3
Pygments==2.17.2
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.3.7
sphinx-basic-ng==1.0.0b2
sphinx-copybutton==0.5.2
sphinx_mdinclude==0.5.4
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-bibtex==2.6.2
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-httpdomain==1.8.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-openapi==0.8.4
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
urllib3==2.2.1
